# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Protocolo Cirúrgico Ortognático — Clínica da Face
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protocols:
    get:
      operationId: listProtocols
      tags: [protocols]
      summary: List all protocols
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: surgeryType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of protocols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProtocolSummary"
    post:
      operationId: createProtocol
      tags: [protocols]
      summary: Create a new protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
...
  /protocols/{id}:
    get:
      operationId: getProtocol
      tags: [protocols]
      summary: Get a protocol by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProtocol
      tags: [protocols]
      summary: Update a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProtocol
      tags: [protocols]
      summary: Delete a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /protocols/{id}/duplicate:
    post:
      operationId: duplicateProtocol
      tags: [protocols]
      summary: Duplicate a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /protocols/{id}/generate-description:
    post:
      operationId: generateOperativeDescription
      tags: [protocols]
      summary: Auto-generate operative description from protocol data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Protocol schemas ───────────────────────────────────────
    SurgicalTeam:
      type: object
      properties:
        surgeon: { type: string }
        firstAssistant: { type: string }
        secondAssistant: { type: string }
        instrumentist: { type: string }
        anesthesiologist: { type: string }
        scrubNurse: { type: string }
        perfusionist: { type: string }
    ChecklistItem:
      type: object
      required: [item, status]
      properties:
        item: { type: string }
        status:
          type: string
          enum: [ok, missing, na]
        notes: { type: string }
    PreopDiagnosis:
      type: object
      properties:
        skeletalClass:
          type: string
          enum: [I, II, III, ""]
        verticalPattern:
          type: string
          enum: [normodivergent, hyperdivergent, hypodivergent, ""]
        maxillaryPosition: { type: string }
        mandibularPosition: { type: string }
        facialAsymmetry: { type: boolean }
        asymmetryDetails: { type: string }
        openBite: { type: boolean }
        crossBite: { type: boolean }
        crowding: { type: boolean }
        overbite: { type: number }
        overjet: { type: number }
        airwayCompromise: { type: boolean }
        tmjSymptoms: { type: boolean }
        additionalNotes: { type: string }
...
    MaxillaPlan:
      type: object
      properties:
        included: { type: boolean }
        osteotomyType:
          type: string
          enum: [LeFort_I, LeFort_II, LeFort_III, segmented, expansion, SARPE, ""]
        segments:
          type: array
          items:
            $ref: "#/components/schemas/MaxillaSegment"
        bonGraft: { type: boolean }
        graftSource: { type: string }
        fixation: { type: string }
        notes: { type: string }
    MandiblePlan:
      type: object
      properties:
        included: { type: boolean }
        osteotomyType:
          type: string
          enum: [BSSO, vertical_ramus, intraoral_vertical_ramus, genioplasty_only, distraction, ""]
        movements:
          $ref: "#/components/schemas/OrthoMovements"
        rigidFixation: { type: boolean }
        fixationType: { type: string }
        condylarPositioning: { type: string }
        notes: { type: string }
    ChinPlan:
      type: object
      properties:
        included: { type: boolean }
        procedure:
          type: string
          enum: [advancement, setback, vertical_reduction, vertical_augmentation, asymmetry_correction, ""]
        movements:
          $ref: "#/components/schemas/OrthoMovements"
        notes: { type: string }
...
    SurgicalPlan:
      type: object
      properties:
        maxilla:
          $ref: "#/components/schemas/MaxillaPlan"
        mandible:
          $ref: "#/components/schemas/MandiblePlan"
        chin:
          $ref: "#/components/schemas/ChinPlan"
        associated:
          type: array
          items:
            $ref: "#/components/schemas/AssociatedProcedure"
    SurgicalSequenceStep:
      type: object
      required: [order, description]
      properties:
        order: { type: integer }
        description: { type: string }
        startTime: { type: ["string", "null"] }
        endTime: { type: ["string", "null"] }
        notes: { type: string }
...
    IntraopRecord:
      type: object
      properties:
        surgeryStartTime: { type: ["string", "null"] }
        surgeryEndTime: { type: ["string", "null"] }
        anesthesiaStartTime: { type: ["string", "null"] }
        anesthesiaEndTime: { type: ["string", "null"] }
        estimatedBloodLoss: { type: ["number", "null"] }
        fluidAdministered: { type: ["number", "null"] }
        complications:
          type: array
          items:
            $ref: "#/components/schemas/IntraopComplication"
        generalNotes: { type: string }
...
    OsteosynthesisMaterials:
      type: object
      properties:
        plates:
          type: array
          items:
            $ref: "#/components/schemas/PlateRecord"
        additionalScrews:
          type: array
          items:
            $ref: "#/components/schemas/ScrewRecord"
        drills:
          type: array
          items:
            $ref: "#/components/schemas/DrillRecord"
        saws:
          type: array
          items:
            $ref: "#/components/schemas/SawRecord"
        otherMaterials:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              quantity: { type: integer }
              notes: { type: string }
    PiezoEquipment:
      type: object
      properties:
        brand:
          type: string
          enum: [wh, mectron, outro, nao_utilizado, ""]
        model: { type: string }
        tip: { type: string }
        serial: { type: string }
        notes: { type: string }
    ProtocolSummary:
      type: object
      required: [id, processNumber, patientName, surgeryType, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        processNumber: { type: string }
        patientName: { type: string }
        patientDOB: { type: ["string", "null"] }
        surgeryDate: { type: ["string", "null"] }
        surgeryType: { type: string }
        status:
          type: string
          enum: [draft, preop_complete, intraop_complete, finalized]
        surgeon: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    Protocol:
      type: object
      required: [id, processNumber, patientName, surgeryType, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        processNumber: { type: string }
        patientName: { type: string }
        patientDOB: { type: ["string", "null"] }
        patientAge: { type: ["integer", "null"] }
        patientGender:
          type: ["string", "null"]
          enum: [M, F, null]
        surgeryDate: { type: ["string", "null"] }
        surgeryType: { type: string }
        status:
          type: string
          enum: [draft, preop_complete, intraop_complete, finalized]
        team:
          $ref: "#/components/schemas/SurgicalTeam"
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        preopDiagnosis:
          $ref: "#/components/schemas/PreopDiagnosis"
        surgicalPlan:
          $ref: "#/components/schemas/SurgicalPlan"
        surgicalSequence:
          type: array
          items:
            $ref: "#/components/schemas/SurgicalSequenceStep"
        intraopRecord:
          $ref: "#/components/schemas/IntraopRecord"
        materials:
          $ref: "#/components/schemas/OsteosynthesisMaterials"
        piezoEquipment:
          $ref: "#/components/schemas/PiezoEquipment"
        surgicalDiagrams:
          $ref: "#/components/schemas/SurgicalDiagrams"
        operativeDescription: { type: string }
        postopNotes: { type: string }
        reopenHistory:
          type: array
          items:
            $ref: "#/components/schemas/ReopenEvent"
        createdAt: { type: string }
        updatedAt: { type: string }
    ReopenEvent:
      type: object
      required: [reopenedAt]
      properties:
        reopenedAt: { type: string }
        reopenedBy: { type: ["string", "null"] }
    ProtocolInput:
      type: object
      required: [processNumber, patientName, surgeryType]
      properties:
        processNumber: { type: string }
        patientName: { type: string }
        patientDOB: { type: string }
        patientAge: { type: integer }
        patientGender:
          type: string
          enum: [M, F]
        surgeryDate: { type: string }
        surgeryType: { type: string }
        status:
          type: string
          enum: [draft, preop_complete, intraop_complete, finalized]
        team:
          $ref: "#/components/schemas/SurgicalTeam"
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        preopDiagnosis:
          $ref: "#/components/schemas/PreopDiagnosis"
        surgicalPlan:
          $ref: "#/components/schemas/SurgicalPlan"
        surgicalSequence:
          type: array
          items:
            $ref: "#/components/schemas/SurgicalSequenceStep"
        intraopRecord:
          $ref: "#/components/schemas/IntraopRecord"
        materials:
          $ref: "#/components/schemas/OsteosynthesisMaterials"
        piezoEquipment:
          $ref: "#/components/schemas/PiezoEquipment"
        surgicalDiagrams:
          $ref: "#/components/schemas/SurgicalDiagrams"
        operativeDescription: { type: string }
        postopNotes: { type: string }
    ProtocolUpdate:
      type: object
      properties:
        processNumber: { type: string }
        patientName: { type: string }
        patientDOB: { type: string }
        patientAge: { type: integer }
        patientGender:
          type: string
          enum: [M, F]
        surgeryDate: { type: string }
        surgeryType: { type: string }
        status:
          type: string
          enum: [draft, preop_complete, intraop_complete, finalized]
        team:
          $ref: "#/components/schemas/SurgicalTeam"
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        preopDiagnosis:
          $ref: "#/components/schemas/PreopDiagnosis"
        surgicalPlan:
          $ref: "#/components/schemas/SurgicalPlan"
        surgicalSequence:
          type: array
          items:
            $ref: "#/components/schemas/SurgicalSequenceStep"
        intraopRecord:
          $ref: "#/components/schemas/IntraopRecord"
        materials:
          $ref: "#/components/schemas/OsteosynthesisMaterials"
        piezoEquipment:
          $ref: "#/components/schemas/PiezoEquipment"
        surgicalDiagrams:
          $ref: "#/components/schemas/SurgicalDiagrams"
        operativeDescription: { type: string }
        postopNotes: { type: string }
        # Only honored on the un-finalize transition: names the clinician who
        # reopened a finalized report. Ignored on all other updates.
        reopenedBy: { type: ["string", "null"] }
...
    SurgicalDiagrams:
      type: object
      properties:
        frontal:
          $ref: "#/components/schemas/DiagramAnnotation"
        lateral:
          $ref: "#/components/schemas/DiagramAnnotation"
        occlusion:
          $ref: "#/components/schemas/DiagramAnnotation"
        molar:
          $ref: "#/components/schemas/DiagramAnnotation"
...
    GeneratedDescription:
      type: object
      required: [description]
      properties:
        description: { type: string }

    # ─── Template schemas ───────────────────────────────────────
...